All the words registered in the database are now displayed in more detail and order.
The requirement to always enter "Q" when adding or deleting words to the database has been removed. Now whenever "Q" is entered it returns to the menu.

The complexity due to the fact that the letter "I" and "i" are not equal to each other in Turkish has been corrected.